feat: topic blacklist

+++ app/src/main/java/com/example/c001apk/ui/activity/BlackListActivity.kt
@@ -11,9 +11,11 @@ import android.widget.TextView
 import androidx.appcompat.widget.ThemeUtils
 import androidx.core.graphics.ColorUtils
 import androidx.lifecycle.ViewModelProvider
+import com.example.c001apk.R
 import com.example.c001apk.adapter.HistoryAdapter
 import com.example.c001apk.databinding.ActivityBlackListBinding
 import com.example.c001apk.logic.database.BlackListDatabase
+import com.example.c001apk.logic.database.TopicBlackListDatabase
 import com.example.c001apk.logic.model.SearchHistory
 import com.example.c001apk.ui.fragment.minterface.IOnItemClickListener
 import com.example.c001apk.viewmodel.AppViewModel
@@ -35,6 +37,9 @@ class BlackListActivity : BaseActivity(), IOnItemClickListener {
     private val blackListDao by lazy {
         BlackListDatabase.getDatabase(this@BlackListActivity).blackListDao()
     }
+    private val topicBlackListDao by lazy {
+        TopicBlackListDatabase.getDatabase(this@BlackListActivity).blackListDao()
+    }
 
     @SuppressLint("NotifyDataSetChanged")
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -42,9 +47,15 @@ class BlackListActivity : BaseActivity(), IOnItemClickListener {
         binding = ActivityBlackListBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        viewModel.type = intent.getStringExtra("type").toString()
+
         initView()
-        if (viewModel.historyList.isEmpty())
-            viewModel.getBlackList("blacklist", this)
+        if (viewModel.historyList.isEmpty()) {
+            when (viewModel.type) {
+                "user" -> viewModel.getBlackList("blacklist", this)
+                "topic" -> viewModel.getBlackList("topicBlacklist", this)
+            }
+        }
 
         initButton()
         initEditText()
@@ -77,7 +88,10 @@ class BlackListActivity : BaseActivity(), IOnItemClickListener {
                 setNegativeButton(android.R.string.cancel, null)
                 setPositiveButton(android.R.string.ok) { _, _ ->
                     CoroutineScope(Dispatchers.IO).launch {
-                        blackListDao.deleteAll()
+                        when (viewModel.type) {
+                            "user" -> blackListDao.deleteAll()
+                            "topic" -> topicBlackListDao.deleteAll()
+                        }
                     }
                     viewModel.historyList.clear()
                     mAdapter.notifyDataSetChanged()
@@ -102,12 +116,22 @@ class BlackListActivity : BaseActivity(), IOnItemClickListener {
 
     @SuppressLint("RestrictedApi")
     private fun initEditText() {
+        binding.title.text = when (viewModel.type) {
+            "user" -> this.getString(R.string.user_black_list)
+            "topic" -> this.getString(R.string.topic_black_list)
+            else -> ""
+        }
         binding.editText.highlightColor = ColorUtils.setAlphaComponent(
             ThemeUtils.getThemeAttrColor(
                 this,
                 rikka.preference.simplemenu.R.attr.colorPrimaryDark
             ), 128
         )
+        binding.editText.hint = when (viewModel.type) {
+            "user" -> "uid"
+            "topic" -> "话题"
+            else -> ""
+        }
         binding.editText.isFocusable = true
         binding.editText.isFocusableInTouchMode = true
         binding.editText.requestFocus()
@@ -115,7 +139,11 @@ class BlackListActivity : BaseActivity(), IOnItemClickListener {
             this.getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
         imm.showSoftInput(binding.editText, 0)
         binding.editText.imeOptions = EditorInfo.IME_ACTION_SEARCH
-        binding.editText.inputType = EditorInfo.TYPE_CLASS_NUMBER
+        binding.editText.inputType = when (viewModel.type) {
+            "user" -> EditorInfo.TYPE_CLASS_NUMBER
+            "topic" -> EditorInfo.TYPE_CLASS_TEXT
+            else -> EditorInfo.TYPE_CLASS_TEXT
+        }
     }
 
     private fun initEdit() {
@@ -143,12 +171,26 @@ class BlackListActivity : BaseActivity(), IOnItemClickListener {
     @SuppressLint("NotifyDataSetChanged")
     private fun updateUid(uid: String) {
         CoroutineScope(Dispatchers.IO).launch {
-            if (blackListDao.isExist(uid)) {
-                viewModel.historyList.remove(uid)
-                blackListDao.delete(uid)
+            when (viewModel.type) {
+                "user" -> {
+                    if (blackListDao.isExist(uid)) {
+                        viewModel.historyList.remove(uid)
+                        blackListDao.delete(uid)
+                    }
+                }
+
+                "topic" -> {
+                    if (topicBlackListDao.isExist(uid)) {
+                        viewModel.historyList.remove(uid)
+                        topicBlackListDao.delete(uid)
+                    }
+                }
             }
             viewModel.historyList.add(0, uid)
-            blackListDao.insert(SearchHistory(uid))
+            when (viewModel.type) {
+                "user" -> blackListDao.insert(SearchHistory(uid))
+                "topic" -> topicBlackListDao.insert(SearchHistory(uid))
+            }
             withContext(Dispatchers.Main) {
                 mAdapter.notifyDataSetChanged()
                 if (binding.clearAll.visibility != View.VISIBLE)
@@ -166,7 +208,10 @@ class BlackListActivity : BaseActivity(), IOnItemClickListener {
     @SuppressLint("NotifyDataSetChanged")
     override fun onItemDeleteClick(keyword: String) {
         CoroutineScope(Dispatchers.IO).launch {
-            blackListDao.delete(keyword)
+            when (viewModel.type) {
+                "user" -> blackListDao.delete(keyword)
+                "topic" -> topicBlackListDao.delete(keyword)
+            }
         }
         viewModel.historyList.remove(keyword)
         mAdapter.notifyDataSetChanged()
